Debug中...(目前已经成功运行前7章所有的测试代码)

+++ tool/stack/stack.c
@@ -99,8 +99,8 @@ uint32 popU32(stack *s) {
     return POP(s, uint32);
 }
 
-uint32 popS64(stack *s) {
-    return POP(s, uint32);
+int64 popS64(stack *s) {
+    return POP(s, int64);
 }
 
 uint64 popU64(stack *s) {
